+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 * Fix for RLMObject being treated as a model object class and showing up in the browser.
 * Fix compilation from the podspec.
+* Fix for crash when calling `objectsWhere:` with grouping in the query on `allObjects`.
 
 0.90.0 Release notes (2015-01-21)
 =============================================================